+++ R/LearnerRegrXgboost.R
@@ -62,6 +62,9 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
           if (is.null(param_vals$early_stopping_rounds)) {
             stop("Parameter 'early_stopping_rounds' must be set to use internal tuning.")
           }
+          if (is.null(param_vals$eval_metric)) {
+            stop("Parameter 'eval_metric' must be set explicitly when using internal tuning.")
+          }
           assert_integerish(domain$upper, len = 1L, any.missing = FALSE) }, .parent = topenv()),
         disable_in_tune = list(early_stopping_rounds = NULL)
       )
@@ -237,7 +240,7 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
       pars = self$param_set$get_values(tags = "train")
       pars_train = self$state$param_vals
       if (!is.null(pars_train$early_stopping_rounds)) {
-        stop("The parameter `early_stopping_rounds` is set. Early stopping and hotstarting are incompatible.")
+        stopf("The parameter `early_stopping_rounds` is set. Early stopping and hotstarting are incompatible.")
       }
 
       # Calculate additional boosting iterations