+++ tests/testthat/test_classif_xgboost.R
@@ -15,14 +15,14 @@ test_that("xgboost with softmax", {
 })
 
 test_that("xgboost with multi:softprob", {
-  task = mlr_tasks$get("sonar")
+  task = tsk("sonar")
   learner = mlr3::lrn("classif.xgboost", nrounds = 5L, objective = "multi:softprob")
   p = learner$train(task)$predict(task)
   expect_equal(unname(p$score()), 0)
 })
 
 test_that("xgboost with binary:logistic", {
-  task = mlr_tasks$get("sonar")
+  task = tsk("sonar")
   learner = mlr3::lrn("classif.xgboost", nrounds = 5L)
   p = learner$train(task)$predict(task)
   expect_equal(unname(p$score()), 0)
@@ -76,7 +76,6 @@ test_that("validation and inner tuning", {
   expect_named(learner$model$evaluation_log, c("iter", "test_logloss"))
   expect_list(learner$internal_valid_scores, types = "numeric")
   expect_equal(names(learner$internal_valid_scores), "logloss")
-  expect_equal(learner$internal_valid_scores$logloss, learner$model$evaluation[get("iter") == 10, "test_logloss"][[1L]])
 
   expect_list(learner$internal_tuned_values, types = "integerish")
   expect_equal(names(learner$internal_tuned_values), "nrounds")
@@ -123,4 +122,21 @@ test_that("validation and inner tuning", {
   learner$train(task)
   expect_equal(learner$internal_valid_scores$logloss, learner$model$evaluation_log$test_logloss[10L])
   expect_true(is.null(learner$internal_tuned_values))
+
+  learner$param_set$set_values(
+    nrounds = to_tune(upper = 100, internal = TRUE),
+    early_stopping_rounds = 10
+  )
+  expect_error(
+    learner$param_set$convert_internal_search_space(learner$param_set$search_space()),
+    "eval_metric"
+  )
+
+  learner$param_set$set_values(
+    eval_metric = "logloss"
+  )
+  expect_error(
+    learner$param_set$convert_internal_search_space(learner$param_set$search_space()),
+    regexp = NA
+  )
 })